update file room service post func

+++ backend/room_service/src/main/java/com/example/room_service/dto/request/AssetRequest.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AssetRequest {
-    int conditionId;
+    long conditionId;
     String assetType;
     String assetName;
     LocalDateTime ownerShipDate;
