+++ backend/room_service/src/main/java/com/example/room_service/dto/request/RoomRequest.java
@@ -13,7 +13,7 @@ public class RoomRequest {
     double floorArea;
     double rentalPrice;
     int maximumOccupancy;
-    int roomTypeId;
+    long roomTypeId;
     boolean roomStatus;
     List<AssetRequest> facilities;
 }
